// Madrasaty.Front/src/app/models/course.ts
export class Course {
        Id: number;
        Name: string;
        DisciplineLevelId: number;
        DisciplineId: number;
        TeacherId: number;
        StartDate: Date;
        EndTime: Date ;
        Begin: Date;
        End: Date;
        ClassroomId: number;
        Discipline: string;
        Classroom: string;
        Teacher: string;
        DisciplineLevel: string;
        Subject: string;
        Jour :string;
        T1 :Date;
        T2 :Date;
        T3:Date;
        RecurrenceId :number;
        DisabledToCreate:Boolean;

}